plugins: migrate the remaining large surface modules to the `props` mapper
Covers support, assistant, inbox, space and devtools — the five modules with the
most surfaces. Their hook-bearing arrows move to a per-plugin surfaces file.

Two consolidations fall out of naming the components: devtools had eleven copies
of "resolve the active space or render nothing", now one `ActiveSpacePanel` that
takes the panel as a prop; plugin-space's three form inputs share one props type.

plugin-space's hue and icon fields called `useCallback` after an early return;
naming them made the conditional hook order visible, so the hooks are hoisted
above the guard.

This is the last of the 281 inline arrows; every `react-surface` module is now
`.ts` with no JSX.
