+++ packages/plugins/plugin-devtools/src/capabilities/react-surface.ts
@@ -3,12 +3,10 @@
 //
 
 import * as Effect from 'effect/Effect';
-import React, { useCallback } from 'react';
 
 import { Capabilities, Capability } from '@dxos/app-framework';
-import { Surface, useOperationInvoker } from '@dxos/app-framework/ui';
-import { GraphPath, LayoutOperation } from '@dxos/app-toolkit';
-import { AppSurface, useActiveSpace } from '@dxos/app-toolkit/ui';
+import { Surface } from '@dxos/app-framework/ui';
+import { AppSurface } from '@dxos/app-toolkit/ui';
 import { EDGE_SERVICE_DEFAULTS, EdgeServiceName } from '@dxos/config';
 import {
   AutomergePanel,
@@ -19,7 +17,6 @@ import {
   EdgeDashboardPanel,
   FeedsPanel,
   IdentityPanel,
-  InvocationTraceContainer,
   KeyringPanel,
   LoggingPanel,
   MembersPanel,
@@ -29,24 +26,24 @@ import {
   QueuesPanel,
   SchemaPanel,
   SignalPanel,
-  SpaceInfoPanel,
-  SpaceListPanel,
   SqlitePanel,
   StoragePanel,
   SwarmPanel,
-  TestingPanel,
   WorkflowPanel,
 } from '@dxos/devtools';
-import { Feed } from '@dxos/echo';
-import { log } from '@dxos/log';
 import { type Graph } from '@dxos/plugin-graph';
-import { ScriptOperation } from '@dxos/plugin-script';
-import { SpaceOperation } from '@dxos/plugin-space';
-import { type Space } from '@dxos/react-client/echo';
 import { ToolsExplorer } from '@dxos/react-ui-introspect';
 
 import { DebugGraph, DevtoolsOverviewContainer, RegistryPanel } from '#containers';
 import { Devtools } from '#types';
+
+import {
+  ActiveSpacePanel,
+  EdgeTestingSurface,
+  EdgeTracesSurface,
+  SpaceInfoSurface,
+  SpaceListSurface,
+} from './DevtoolsSurfaces';
 
 const MCP_SERVER_URL = EDGE_SERVICE_DEFAULTS[EdgeServiceName.Introspect];
 
@@ -68,22 +65,24 @@ export default Capability.makeModule(
       Surface.create({
         id: 'appGraph',
         filter: AppSurface.subject(AppSurface.Article, isGraphDebug),
-        component: ({ data }) => <DebugGraph graph={data.subject.graph} root={data.subject.root} />,
+        component: DebugGraph,
+        props: ({ data: { subject } }) => ({ graph: subject.graph, root: subject.root }),
       }),
       Surface.create({
         id: 'toolsExplorer',
         filter: AppSurface.literal(AppSurface.Article, Devtools.ToolsExplorer),
-        component: () => <ToolsExplorer serverUrl={MCP_SERVER_URL} />,
+        component: ToolsExplorer,
+        props: () => ({ serverUrl: MCP_SERVER_URL }),
       }),
       Surface.create({
         id: 'registry',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Registry),
-        component: () => <RegistryPanel />,
+        component: RegistryPanel,
       }),
       Surface.create({
         id: 'devtoolsOverview',
         filter: Surface.makeFilter(AppSurface.deckCompanion('devtoolsOverview')),
-        component: () => <DevtoolsOverviewContainer />,
+        component: DevtoolsOverviewContainer,
       }),
 
       //
@@ -93,238 +92,136 @@ export default Capability.makeModule(
       Surface.create({
         id: 'client.config',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Client.Config),
-        component: () => <ConfigPanel vaultSelector={false} />,
+        component: ConfigPanel,
+        props: () => ({ vaultSelector: false }),
       }),
       Surface.create({
         id: 'client.storage',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Client.Storage),
-        component: () => <StoragePanel />,
+        component: StoragePanel,
       }),
       Surface.create({
         id: 'client.sqlite',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Client.Sqlite),
-        component: () => <SqlitePanel />,
+        component: SqlitePanel,
       }),
       Surface.create({
         id: 'client.logs',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Client.Logs),
-        component: () => <LoggingPanel />,
+        component: LoggingPanel,
       }),
       Surface.create({
         id: 'client.diagnostics',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Client.Diagnostics),
-        component: () => <DiagnosticsPanel />,
+        component: DiagnosticsPanel,
       }),
       Surface.create({
         id: 'halo.identity',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Halo.Identity),
-        component: () => <IdentityPanel />,
+        component: IdentityPanel,
       }),
       Surface.create({
         id: 'halo.devices',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Halo.Devices),
-        component: () => <DeviceListPanel />,
+        component: DeviceListPanel,
       }),
       Surface.create({
         id: 'halo.keyring',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Halo.Keyring),
-        component: () => <KeyringPanel />,
+        component: KeyringPanel,
       }),
       Surface.create({
         id: 'halo.credentials',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Halo.Credentials),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <CredentialsPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: CredentialsPanel }),
       }),
       Surface.create({
         id: 'echo.spaces',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Spaces),
-        component: () => {
-          const { invokePromise } = useOperationInvoker();
-          const handleSelect = useCallback(
-            () => invokePromise(LayoutOperation.Open, { subject: [Devtools.Echo.Space] }),
-            [invokePromise],
-          );
-          return <SpaceListPanel onSelect={handleSelect} />;
-        },
+        component: SpaceListSurface,
       }),
       Surface.create({
         id: 'echo.space',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Space),
-        component: () => {
-          const space = useActiveSpace();
-          const { invokePromise } = useOperationInvoker();
-          const handleSelect = useCallback(
-            () => invokePromise(LayoutOperation.Open, { subject: [Devtools.Echo.Feeds] }),
-            [invokePromise],
-          );
-          if (!space) {
-            return null;
-          }
-
-          return <SpaceInfoPanel space={space} onSelectFeed={handleSelect} onSelectPipeline={handleSelect} />;
-        },
+        component: SpaceInfoSurface,
       }),
       Surface.create({
         id: 'echo.feeds',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Feeds),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <FeedsPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: FeedsPanel }),
       }),
       Surface.create({
         id: 'echo.objects',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Objects),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <ObjectsPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: ObjectsPanel }),
       }),
       Surface.create({
         id: 'echo.schema',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Schema),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <SchemaPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: SchemaPanel }),
       }),
       Surface.create({
         id: 'echo.automerge',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Automerge),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <AutomergePanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: AutomergePanel }),
       }),
       Surface.create({
         id: 'echo.queues',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Queues),
-        component: () => <QueuesPanel />,
+        component: QueuesPanel,
       }),
       Surface.create({
         id: 'echo.members',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Members),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <MembersPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: MembersPanel }),
       }),
       Surface.create({
         id: 'echo.metadata',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Echo.Metadata),
-        component: () => <MetadataPanel />,
+        component: MetadataPanel,
       }),
       Surface.create({
         id: 'mesh.signal',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Mesh.Signal),
-        component: () => <SignalPanel />,
+        component: SignalPanel,
       }),
       Surface.create({
         id: 'mesh.swarm',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Mesh.Swarm),
-        component: () => <SwarmPanel />,
+        component: SwarmPanel,
       }),
       Surface.create({
         id: 'mesh.network',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Mesh.Network),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <NetworkPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: NetworkPanel }),
       }),
       Surface.create({
         id: 'edge.dashboard',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Edge.Dashboard),
-        component: () => <EdgeDashboardPanel />,
+        component: EdgeDashboardPanel,
       }),
       Surface.create({
         id: 'edge.workflows',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Edge.Workflows),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          return <WorkflowPanel space={space} />;
-        },
+        component: ActiveSpacePanel,
+        props: () => ({ Panel: WorkflowPanel }),
       }),
       Surface.create({
         id: 'edge.traces',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Edge.Traces),
-        component: () => {
-          const space = useActiveSpace();
-          if (!space) {
-            return null;
-          }
-
-          const feed = space.properties.invocationTraceFeed?.target;
-          const feedDXN = feed ? Feed.getFeedUri(feed) : undefined;
-          return <InvocationTraceContainer db={space.db} feedDXN={feedDXN} detailAxis='block' />;
-        },
+        component: EdgeTracesSurface,
       }),
       Surface.create({
         id: 'edge.testing',
         filter: AppSurface.literal(AppSurface.Article, Devtools.Edge.Testing),
-        component: () => {
-          const { invokePromise } = useOperationInvoker();
-          const onSpaceCreate = useCallback(
-            async (space: Space) => {
-              await space.waitUntilReady();
-              await invokePromise(SpaceOperation.Migrate, { space });
-              await space.db.flush();
-            },
-            [invokePromise],
-          );
-          const onScriptPluginOpen = useCallback(
-            async (space: Space) => {
-              await space.waitUntilReady();
-              const createResult = await invokePromise(ScriptOperation.CreateScript, { db: space.db });
-              if (createResult.data?.object) {
-                await invokePromise(SpaceOperation.AddObject, { target: space.db, object: createResult.data.object });
-              }
-              log.info('script created', { result: createResult });
-              if (createResult.data?.object) {
-                await invokePromise(LayoutOperation.Open, {
-                  subject: [GraphPath.getObjectPathFromObject(createResult.data.object)],
-                });
-              }
-            },
-            [invokePromise],
-          );
-          return <TestingPanel onSpaceCreate={onSpaceCreate} onScriptPluginOpen={onScriptPluginOpen} />;
-        },
+        component: EdgeTestingSurface,
       }),
     ]);
   }),
